refactor: menu bar
