+++ src/scss/index.scss
@@ -15,6 +15,7 @@
 @import 'parts/comment-card';
 @import 'parts/code-and-comment-selector';
 @import 'parts/repository-selector';
+@import 'parts/menu-bar';
 
 html, body {
   font-family: Roboto, sans-serif;
